+++ MovieShop/Models/ShoppingCartMovie.cs
@@ -10,6 +10,6 @@
 
         public int Quantity { get; set; } = 1;
 
-        public decimal PriceAtOrder { get; set; }
+        public int PriceAtOrder { get; set; }
     }
 }
